+++ CHANGES.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+ - Add the --smt-lib-fpa cmdline option to activate support for the SMT-LIB FPA
+   theory and support parsing of SMT-LIB FPA Theory literals (#1358)
  - Bump minimal cmdliner version to 2.0 (#1340)
 
 ## v2.6.2